add DesignModelFwd.hpp
